fixed remote method semantics to return owned entities instead of accessible entities allow deletion with write permission apply permissions recursively by default
Former-commit-id: 044c52f24abee1525088e149d1e0777027a2f2e9

+++ console/src/main/java/org/janelia/it/FlyWorkstation/gui/framework/outline/EntityContextMenu.java
@@ -806,7 +806,7 @@ public class EntityContextMenu extends JPopupMenu {
 
         for (RootedEntity rootedEntity : rootedEntityList) {
             EntityData ed = rootedEntity.getEntityData();
-            if (ed.getId() == null && EntityUtils.isCommonRoot(ed.getChildEntity())) {
+            if (ed.getId() == null && !EntityUtils.isCommonRoot(ed.getChildEntity())) {
                 // Fake ED, not a common root, this must be part of an
                 // annotation session.
                 // TODO: this check could be done more robustly
@@ -822,31 +822,32 @@ public class EntityContextMenu extends JPopupMenu {
                 action.doAction();
             }
         });
-
+        
         for (RootedEntity rootedEntity : rootedEntityList) {
             Entity entity = rootedEntity.getEntity();
             Entity parent = rootedEntity.getEntityData().getParentEntity();
-            if (ModelMgrUtils.hasWriteAccess(entity)) {
-                // User can delete because they own this entity
-                deleteItem.setEnabled(true);
-                // Unless it is protected
-                if (EntityUtils.isProtected(entity)) {
-                    deleteItem.setEnabled(false);
-                }
-                // Or its parent is protected
-                if (parent!=null && parent.getId()!=null && EntityUtils.isProtected(parent)) {
-                    deleteItem.setEnabled(false);
-                }
-            }
-            else {
-                deleteItem.setEnabled(false);
-                // User can't delete unless they own the parent
+            
+            boolean canDelete = true;
+            // User can't delete if they don't have write access
+            if (!ModelMgrUtils.hasWriteAccess(entity)) {
+                canDelete = false;
+                // Unless they own the parent
                 if (parent!=null && parent.getId()!=null && ModelMgrUtils.hasWriteAccess(parent)) {
-                    deleteItem.setEnabled(true);
+                    canDelete = true;
                 }
             }
+            // Can never delete protected entities
+            if (EntityUtils.isProtected(entity)) {
+                canDelete = false;
+            }
+            // Or children of protected entities
+            if (parent!=null && parent.getId()!=null && EntityUtils.isProtected(parent)) {
+                canDelete = false;
+            }
+            
+            if (!canDelete) deleteItem.setEnabled(false);
         }
-
+        
         return deleteItem;
     }
